simplify implementation of GeometryTrait on Unimplemented*
