+++ geo-traits/src/multi_point.rs
@@ -1,12 +1,7 @@
 use std::marker::PhantomData;
 
 use crate::iterator::MultiPointIterator;
-use crate::{
-    Dimensions, GeometryTrait, GeometryType, PointTrait, UnimplementedGeometryCollection,
-    UnimplementedLine, UnimplementedLineString, UnimplementedMultiLineString,
-    UnimplementedMultiPolygon, UnimplementedPoint, UnimplementedPolygon, UnimplementedRect,
-    UnimplementedTriangle,
-};
+use crate::{GeometryTrait, PointTrait, UnimplementedPoint};
 #[cfg(feature = "geo-types")]
 use geo_types::{CoordNum, MultiPoint, Point};
 
@@ -99,71 +94,5 @@ impl<T> MultiPointTrait for UnimplementedMultiPoint<T> {
 
     unsafe fn point_unchecked(&self, _i: usize) -> Self::InnerPointType<'_> {
         unimplemented!()
-    }
-}
-
-impl<T> GeometryTrait for UnimplementedMultiPoint<T> {
-    type T = T;
-    type PointType<'b>
-        = UnimplementedPoint<Self::T>
-    where
-        Self: 'b;
-    type LineStringType<'b>
-        = UnimplementedLineString<Self::T>
-    where
-        Self: 'b;
-    type PolygonType<'b>
-        = UnimplementedPolygon<Self::T>
-    where
-        Self: 'b;
-    type MultiPointType<'b>
-        = UnimplementedMultiPoint<Self::T>
-    where
-        Self: 'b;
-    type MultiLineStringType<'b>
-        = UnimplementedMultiLineString<Self::T>
-    where
-        Self: 'b;
-    type MultiPolygonType<'b>
-        = UnimplementedMultiPolygon<Self::T>
-    where
-        Self: 'b;
-    type GeometryCollectionType<'b>
-        = UnimplementedGeometryCollection<Self::T>
-    where
-        Self: 'b;
-    type RectType<'b>
-        = UnimplementedRect<Self::T>
-    where
-        Self: 'b;
-    type TriangleType<'b>
-        = UnimplementedTriangle<Self::T>
-    where
-        Self: 'b;
-    type LineType<'b>
-        = UnimplementedLine<Self::T>
-    where
-        Self: 'b;
-
-    fn dim(&self) -> Dimensions {
-        unimplemented!()
-    }
-
-    fn as_type(
-        &self,
-    ) -> GeometryType<
-        '_,
-        Self::PointType<'_>,
-        Self::LineStringType<'_>,
-        Self::PolygonType<'_>,
-        Self::MultiPointType<'_>,
-        Self::MultiLineStringType<'_>,
-        Self::MultiPolygonType<'_>,
-        Self::GeometryCollectionType<'_>,
-        Self::RectType<'_>,
-        Self::TriangleType<'_>,
-        Self::LineType<'_>,
-    > {
-        GeometryType::MultiPoint(self)
     }
 }
